+++ src/crystal/tests/index.py
@@ -8,7 +8,7 @@ from crystal.tests import (
     test_icons, test_install_to_desktop, test_load_urls, test_log_drawer,
     test_menus, test_new_group, test_new_root_url, test_open_project,
     test_parse_html, test_profile, test_project_migrate, test_readonly_mode,
-    test_server, test_shell, test_ssd, test_tasks, test_tasktree,
+    test_runner, test_server, test_shell, test_ssd, test_tasks, test_tasktree,
     test_untitled_projects,
     test_workflows, test_xthreading,
 )
@@ -68,6 +68,7 @@ _TEST_FUNCS = (
     _test_functions_in_module(test_profile) +
     _test_functions_in_module(test_project_migrate) +
     _test_functions_in_module(test_readonly_mode) +
+    _test_functions_in_module(test_runner) +
     _test_functions_in_module(test_server) +
     _test_functions_in_module(test_shell) +
     _test_functions_in_module(test_ssd) +